don't use SoftmaxActionSampler (#354)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/354

`Categorical.sample()` keeps throwing exception.

Reviewed By: kaiwenw

Differential Revision: D25469106

fbshipit-source-id: 4ee12ae9185b5e60b242e4c0edae6a9d91c2c145

+++ reagent/gym/policies/samplers/discrete_sampler.py
@@ -76,7 +76,9 @@ class GreedyActionSampler(Sampler):
         raw_action = self._get_greedy_indices(scores)
         action = F.one_hot(raw_action, num_actions)
         assert action.shape == (batch_size, num_actions)
-        return rlt.ActorOutput(action=action, log_prob=torch.ones_like(raw_action))
+        return rlt.ActorOutput(
+            action=action, log_prob=torch.ones_like(raw_action, dtype=torch.float)
+        )
 
     # pyre-fixme[56]: Decorator `torch.no_grad(...)` could not be called, because
     #  its type `no_grad` is not callable.
